accesproject/requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.12.4
mssql-django==1.4.2
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.4.4
